+++ public_html/admin-scenarios.php
@@ -81,7 +81,7 @@
 						<li class="list-group-item d-flex flex-wrap justify-content-between">
 							<?php echo $scenario["name"]." (Created On ".date("d M, Y", strtotime($scenario["date"]))." at ".date("H:i", strtotime($scenario["date"]))." Z)"; ?> 
 							<div id="buttons" class="d-inline">
-								<a class="btn btn-success" href="/cas?scenario=<?php echo $scenario["id"];?>" role="button">Load</a>
+								<a class="btn btn-success" href="/cas?scenario[]=<?php echo $scenario["id"];?>" role="button">Load</a>
 								<button type="button" class="btn btn-primary btn-share-scenario" data-name="<?php echo $scenario["name"]; ?>" data-id="<?php echo $scenario["id"]; ?>" data-toggle="modal" data-target="#share-scenario-modal">Share</button>
 								<button type="button" class="btn btn-danger btn-del" data-name="<?php echo $scenario["name"]; ?>" data-id="<?php echo $scenario["id"]; ?>" data-toggle="modal" data-target="#del-scenario-modal">Delete</button>
 							</div>